Update package manager and dependencies (#52)
