+++ README.md
@@ -238,26 +238,25 @@ git push origin master
 
 ### How to Check by Running the Server in Local
 
-You don't need to do this process, but if you're interested, here is the process.
-
-Note: worked with Node v16, but got error with Node v18. (should be fixed in the future)
-
 ```shell
 git clone git@github.com:ucalgary-ilab/ilab-website.git
 cd ilab-website
-npm install
-npm run build:content
+yarn install
+yarn build:content
 node preprocess.js
-node server.js
+yarn dev
 ```
 
-You can also keep watching the file change by running the following command in the different terminal tab
+Then browse: http://localhost:3000
 
-```shell
-npm run watch
-```
+Notes: 
+- `yarn build:content` converts files under [content](content) from Markdown and YAML to JSON. You can monitor Markdown and YAML files changes and automate their conversion to JSON files by running the following command in another terminal tab:
+  ```shell
+  yarn watch
+  ```
+- `node preprocess.js` needs to be run every time image files are added under `static`.
 
-Watch [https://youtu.be/ympcMWQHm1c](https://youtu.be/ympcMWQHm1c) for the video instruction.
+(equivalent to `yarn build:content`)
 
 ### How to Deploy and Reflect the Change in the Website
 
@@ -276,5 +275,5 @@ git clone git@github.com:ucalgary-ilab/ucalgary-ilab.github.io.git -b gh-pages .
 #### Deploy
 
 ```shell
-npm run deploy
+yarn deploy
 ```
